Add Password Reuse Policy feature
All users passwords are historicized in table (credcheck.pg_auth_history)
together with the timestamps of when these passwords were set.

Two settings allow to control the behavior of this feature:

  - credcheck.password_reuse_history: number of distinct passwords set
    before a password can be reused.
  - credcheck.password_reuse_interval: amount of time it takes before a
    password can be reused again.

The default value for these settings are 0 which means that all password
reuse policies are disabled.

The password history consists of passwords a user has been assigned in
the past. credcheck can restrict new passwords from being chosen from
this history:

  - If an account is restricted on the basis of number of password changes,
    a new password cannot be chosen from the password_reuse_history most
    recent passwords. For example, minimum number of password changes is
    set to 3, a new password cannot be the same as any of the most recent
    3 passwords.

  - If an account is restricted based on time elapsed, a new password can't
    be chosen from those in the history that are newer than the number of
    day set to password_reuse_interval. For example, if the password reuse
    interval is set to 365, new password must not be among those previously
    chosen within the last year.

Thanks to Umair Shahid and Gabi201265 for the feature request.