Add the possibility to check the easiness of a password by the use of the cracklib tool.
If you want to use the "deny password that can be easily cracked" feature you
need to edit the `Makefile` to enable the following lines:

  #PG_CPPFLAGS = -DUSE_CRACKLIB '-DCRACKLIB_DICTPATH="/usr/lib/cracklib_dict"'
  #SHLIB_LINK = -lcrack

Following your distribution the cracklib dictionary might not be at the same place.
For RedHat, CentOs and other rpm like distribution you should use

  PG_CPPFLAGS = -DUSE_CRACKLIB '-DCRACKLIB_DICTPATH="/usr/lib/cracklib_dict"'

and for Debian, Ubuntu and other debian like distribution it should be:

   PG_CPPFLAGS = -DUSE_CRACKLIB '-DCRACKLIB_DICTPATH="/usr/share/dict/cracklib-small"'

This work is simply a integration of a copy/paste from the passwordcheck extension
available in the contrib/ directory. Credits to the author Laurenz Albe.